Fixed tapping issue on iOS devices
Added media queries to make :hover effects specific to desktop devices
in order to avoid an issue on iOS where tapping a clickable element
triggers the hover animation instead of the intended click event. Also
removed a comma that was bugging me from the copy on one of the pages,
but shhhhh.